Refactoring: add flag Wall; fix Wall warn;

+++ api.hpp
@@ -8,5 +8,5 @@ namespace api
     using shift = std::string;
     using change_rules = std::unordered_map<pattern, shift>;
     using file_path = std::string;
-    using task_id_t = int64_t;
+    using task_id_t = size_t;
 }
